area light lut binary asset in project settings and handled in application
